+++ Assets/VRM10/Tests/ExpressionTests.cs
@@ -1,4 +1,3 @@
-using System.IO;
 using System.Linq;
 using NUnit.Framework;
 using UnityEngine;
@@ -7,34 +6,10 @@ namespace UniVRM10.Test
 {
     public class ExpressionTests
     {
-        static string AliciaPath
-        {
-            get
-            {
-                return Path.GetFullPath(Application.dataPath + "/../Tests/Models/Alicia_vrm-0.51/AliciaSolid_vrm-0.51.vrm")
-                    .Replace("\\", "/");
-            }
-        }
-
-        static VRM10Controller Load()
-        {
-            Vrm10Data.TryParseOrMigrate(AliciaPath, true, out Vrm10Data vrm);
-            using (var loader = new Vrm10Importer(vrm))
-            {
-                var task = loader.LoadAsync(new VRMShaders.ImmediateCaller());
-                task.Wait();
-
-                var instance = task.Result;
-
-                return instance.GetComponent<VRM10Controller>();
-            }
-
-        }
-
         [Test]
         public void DuplicatedMaterialColorBindings()
         {
-            var controller = Load();
+            var controller = TestAsset.LoadAlicia();
 
             var src = controller.Vrm.Expression.Aa.MaterialColorBindings.ToList();
 
@@ -63,7 +38,7 @@ namespace UniVRM10.Test
         [Test]
         public void DuplicatedMaterialUVBindings()
         {
-            var controller = Load();
+            var controller = TestAsset.LoadAlicia();
 
             var renderers = controller.GetComponentsInChildren<Renderer>();
             var name = renderers[0].sharedMaterials[0].name;
