+++ Assets/VRM10/Tests/LoadTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c9a9a3ce2deae9c4791d816aa189df3d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
