+++ Assets/VRM10/Tests/TestAsset.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b18f57dcd7d3c19469aa4842b2ee6ae1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
